requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
google-genai>=1.0.0
youtube-transcript-api>=0.6.0
python-dotenv>=1.0.0
